Add function and keybinding for using zotero bbtex CAYW
Uses hs.http for the bbtex CAYW API and recursive key bind for the keys

Making a request using hs.http is simple enough.

Main issue is that zotero retains window focus and doesn't know how to
give it back.
This is fixed by first recording which app is front-most when the funct
is called and returning focus with app:activate() at the end.

Citation is then copied to the pasteboard.

Note the recursive key bindings:
- hyper+k (start recursive key bindings)
- z (zotero
- c, p, b (different formats)